Add diagnostics for new device funnel (#1075)
* Add device metrics

Add metrics for new device flyout

* Add device created metric

Add metric on device create so we can log device id for physical devices

* Fix rule apply event

Rule apply click event was only emitted for new rules. Move call so it is emitted if a rule is added or edited.

* Address comments and align metric names

Address comments. Update metric names to be in same format as rule metrics